fixed PIC issue
the issue turned out not to be the PIC nor the IRQs at all.  in fact, my
`putch()' function was not resetting the row to row 24 after scrolling.  this
eventually led to trying to change values in the VGA buffer outside the
juristiction of where i should be changing values, so garbage was being
printed to the screen.